combined request and response classes into one module. Right now they are used inside core as a default for getattr, to check isinstance, and once to turn alexa request into object
