+++ flask_ask/__init__.py
@@ -15,5 +15,4 @@ from .core import (
     convert_errors
 )
 
-# from .request import Request
-from .response import question, statement, audio
+from .models import question, statement, audio
